+++ RELEASE-NOTES.md
@@ -24,6 +24,8 @@ Picocli follows [semantic versioning](http://semver.org/).
 - [#503] Build: Upgrade to gradle 4.10.2.
 - [#497] add module `picocli-shell-jline2` for components and documentation for building interactive shell command line applications with JLine 2 and picocli.
 - [#499] add module `picocli-codegen` for tools to generate documentation, configuration, source code and other files from a picocli model
+- [#410] add `ReflectionConfigGenerator` class for GraalVM `native-image`
+
 
 ## <a name="3.7.0-deprecated"></a> Deprecations
 No features were deprecated in this release.